+++ packages/manager/src/features/Linodes/LinodeCreate/Networking/InterfaceFirewall.tsx
@@ -2,10 +2,14 @@ import { Box, Stack } from '@linode/ui';
 import React, { useState } from 'react';
 import { useController, useFormContext, useWatch } from 'react-hook-form';
 
+import { AkamaiBanner } from 'src/components/AkamaiBanner/AkamaiBanner';
+import { GenerateFirewallDialog } from 'src/components/GenerateFirewallDialog/GenerateFirewallDialog';
 import { LinkButton } from 'src/components/LinkButton';
 import { FirewallSelect } from 'src/features/Firewalls/components/FirewallSelect';
 import { CreateFirewallDrawer } from 'src/features/Firewalls/FirewallLanding/CreateFirewallDrawer';
+import { useFlags } from 'src/hooks/useFlags';
 import { useRestrictedGlobalGrantCheck } from 'src/hooks/useRestrictedGlobalGrantCheck';
+import { useSecureVMNoticesEnabled } from 'src/hooks/useSecureVMNoticesEnabled';
 
 import type { LinodeCreateFormValues } from '../utilities';
 
@@ -14,6 +18,14 @@ interface Props {
 }
 
 export const InterfaceFirewall = ({ index }: Props) => {
+  const { secureVMNoticesEnabled } = useSecureVMNoticesEnabled();
+  const flags = useFlags();
+
+  const [
+    isGenerateAkamaiEmployeeFirewallDialogOpen,
+    setIsGenerateAkamaiEmployeeFirewallDialogOpen,
+  ] = useState(false);
+
   const { control } = useFormContext<LinodeCreateFormValues>();
 
   const interfaceType = useWatch({
@@ -35,6 +47,24 @@ export const InterfaceFirewall = ({ index }: Props) => {
   return (
     <Stack spacing={2}>
       <Stack spacing={1.5}>
+        {secureVMNoticesEnabled && (
+          <AkamaiBanner
+            action={
+              <LinkButton
+                onClick={() =>
+                  setIsGenerateAkamaiEmployeeFirewallDialogOpen(true)
+                }
+              >
+                {flags.secureVmCopy?.generateActionText ??
+                  'Generate Compliant Firewall'}
+              </LinkButton>
+            }
+            text={
+              flags.secureVmCopy?.linodeCreate?.text ??
+              'All accounts must apply an compliant firewall to all their Linodes.'
+            }
+          />
+        )}
         <FirewallSelect
           disabled={isLinodeCreateRestricted}
           errorText={fieldState.error?.message}
@@ -59,6 +89,13 @@ export const InterfaceFirewall = ({ index }: Props) => {
         onFirewallCreated={(firewall) => field.onChange(firewall.id)}
         open={isDrawerOpen}
       />
+      {secureVMNoticesEnabled && (
+        <GenerateFirewallDialog
+          onClose={() => setIsGenerateAkamaiEmployeeFirewallDialogOpen(false)}
+          onFirewallGenerated={(firewall) => field.onChange(firewall.id)}
+          open={isGenerateAkamaiEmployeeFirewallDialogOpen}
+        />
+      )}
     </Stack>
   );
 };